Django/pyproject.toml
[tool.poetry]
name = "oz-my-django-project"
version = "0.1.0"
description = "py djg project"
authors = ["Eojs <jsduek28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pyjwt = "^2.10.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
